+++ Missionframework/KPLIB_functions.hpp
@@ -4,7 +4,7 @@
     File: KPLIB_functions.hpp
     Author: KP Liberation Dev Team - https://github.com/KillahPotatoes
     Date: 2017-10-16
-    Last Update: 2019-04-05
+    Last Update: 2019-04-22
     License: GNU General Public License v3.0 - https://www.gnu.org/licenses/gpl-3.0.html
 
     Description:
@@ -18,7 +18,7 @@ class KPLIB {
     #include "modules\02_core\functions.hpp"
     #include "modules\03_persistence\functions.hpp"
     #include "modules\04_respawn\functions.hpp"
-    #include "modules\05_adm\functions.hpp"
+    #include "modules\05_admin\functions.hpp"
     #include "modules\10_resources\functions.hpp"
     #include "modules\11_permission\functions.hpp"
     #include "modules\12_arsenal\functions.hpp"
