+++ Missionframework/modules/05_admin/README.md
@@ -7,27 +7,27 @@ The Admin Menu provides maintenance and debug functionalities for the logged in 
 * Init
 
 ### Functions
-* KPLIB_fnc_adm_deleteExport
+* KPLIB_fnc_admin_deleteExport
 
   *Deletes exported data in the players profile.*
 
-* KPLIB_fnc_adm_exportSave
+* KPLIB_fnc_admin_exportSave
 
   *Exports current save data to the players profile.*
 
-* KPLIB_fnc_adm_importSave
+* KPLIB_fnc_admin_importSave
 
   *Imports save data from the players profile.*
 
-* KPLIB_fnc_adm_importSaveServer
+* KPLIB_fnc_admin_importSaveServer
 
   *Applies provided save data for import to the server profile.*
 
-* KPLIB_fnc_adm_openDialog
+* KPLIB_fnc_admin_openDialog
 
   *Opens the admin menu dialog.*
 
-* KPLIB_fnc_adm_wipe
+* KPLIB_fnc_admin_wipe
 
   *Export current campaign to admins profile and requests server to wipe the saved campaign.*
