+++ Missionframework/modules/05_admin/fnc/fn_admin_deleteExport.sqf
@@ -1,10 +1,10 @@
 /*
-    KPLIB_fnc_adm_deleteExport
+    KPLIB_fnc_admin_deleteExport
 
-    File: fn_adm_deleteExport.sqf
+    File: fn_admin_deleteExport.sqf
     Author: KP Liberation Dev Team - https://github.com/KillahPotatoes
     Date: 2018-08-02
-    Last Update: 2018-11-10
+    Last Update: 2019-04-22
     License: GNU General Public License v3.0 - https://www.gnu.org/licenses/gpl-3.0.html
 
     Description:
