+++ Missionframework/modules/05_admin/fnc/fn_admin_importSave.sqf
@@ -1,10 +1,10 @@
 /*
-    KPLIB_fnc_adm_importSave
+    KPLIB_fnc_admin_importSave
 
-    File: fn_adm_importSave.sqf
+    File: fn_admin_importSave.sqf
     Author: KP Liberation Dev Team - https://github.com/KillahPotatoes
     Date: 2018-07-27
-    Last Update: 2018-11-10
+    Last Update: 2019-04-22
     License: GNU General Public License v3.0 - https://www.gnu.org/licenses/gpl-3.0.html
 
     Description:
@@ -19,10 +19,10 @@
 
 if (isServer) then {
     // If it's local hosted we can directly import the save data
-    [profileNamespace getVariable [KPLIB_save_key + "_export", nil]] call KPLIB_fnc_adm_importSaveServer;
+    [profileNamespace getVariable [KPLIB_save_key + "_export", nil]] call KPLIB_fnc_admin_importSaveServer;
 } else {
     // Otherwise send data for import to the server
-    [profileNamespace getVariable [KPLIB_save_key + "_export", nil]] remoteExecCall ["KPLIB_fnc_adm_importSaveServer", 2];
+    [profileNamespace getVariable [KPLIB_save_key + "_export", nil]] remoteExecCall ["KPLIB_fnc_admin_importSaveServer", 2];
 };
 
 hint localize "STR_KPLIB_DIALOG_ADMIN_IMP_NOTE";
