+++ Missionframework/modules/05_admin/fnc/fn_admin_openDialog.sqf
@@ -1,10 +1,10 @@
 /*
-    KPLIB_fnc_adm_openDialog
+    KPLIB_fnc_admin_openDialog
 
-    File: fn_adm_openDialog.sqf
+    File: fn_admin_openDialog.sqf
     Author: KP Liberation Dev Team - https://github.com/KillahPotatoes
     Date: 2018-07-27
-    Last Update: 2018-11-10
+    Last Update: 2019-04-22
     License: GNU General Public License v3.0 - https://www.gnu.org/licenses/gpl-3.0.html
 
     Description:
