+++ Missionframework/modules/05_admin/fnc/fn_admin_preInit.sqf
@@ -1,10 +1,10 @@
 /*
-    KPLIB_fnc_adm_preInit
+    KPLIB_fnc_admin_preInit
 
-    File: fn_adm_preInit.sqf
+    File: fn_admin_preInit.sqf
     Author: KP Liberation Dev Team - https://github.com/KillahPotatoes
     Date: 2018-10-18
-    Last Update: 2018-11-12
+    Last Update: 2019-04-22
     License: GNU General Public License v3.0 - https://www.gnu.org/licenses/gpl-3.0.html
 
     Description:
@@ -17,8 +17,8 @@
         Module preInit finished [BOOL]
 */
 
-if (isServer) then {diag_log format ["[KP LIBERATION] [%1] [PRE] [ADM] Module initializing...", diag_tickTime];};
+if (isServer) then {["Module initializing...", "PRE] [ADMIN", true] call KPLIB_fnc_common_log;};
 
-if (isServer) then {diag_log format ["[KP LIBERATION] [%1] [PRE] [ADM] Module initialized", diag_tickTime];};
+if (isServer) then {["Module initialized", "PRE] [ADMIN", true] call KPLIB_fnc_common_log;};
 
 true
