+++ Missionframework/modules/05_admin/functions.hpp
@@ -4,7 +4,7 @@
     File: functions.hpp
     Author: KP Liberation Dev Team - https://github.com/KillahPotatoes
     Date: 2018-08-02
-    Last Update: 2019-03-30
+    Last Update: 2019-04-22
     License: GNU General Public License v3.0 - https://www.gnu.org/licenses/gpl-3.0.html
 
     Description:
@@ -12,33 +12,33 @@
 */
 
 class adm {
-    file = "modules\05_adm\fnc";
+    file = "modules\05_admin\fnc";
 
     // Deletes exported data in the players profile
-    class adm_deleteExport {};
+    class admin_deleteExport {};
 
     // Exports current save data to the players profile
-    class adm_exportSave {};
+    class admin_exportSave {};
 
     // Imports save data from the players profile
-    class adm_importSave {};
+    class admin_importSave {};
 
     // Applies provided save data for import to the server profile
-    class adm_importSaveServer {};
+    class admin_importSaveServer {};
 
     // Opens the admin menu dialog
-    class adm_openDialog {};
+    class admin_openDialog {};
 
     // Module post initialization
-    class adm_postInit {
+    class admin_postInit {
         postInit = 1;
     };
 
     // Module pre initialization
-    class adm_preInit {
+    class admin_preInit {
         preInit = 1;
     };
 
     // Export current campaign to admins profile and requests server to wipe the saved campaign
-    class adm_wipe {};
+    class admin_wipe {};
 };
