+++ Missionframework/modules/05_admin/ui/KPLIB_admin.hpp
@@ -31,7 +31,7 @@ class KPLIB_admin {
             text = "$STR_KPLIB_DIALOG_ADMIN_EXP";
             y = safeZoneY + safeZoneH * (KP_Y_VAL_C + KP_HEIGTH_TITLE + KP_SPACING_Y);
             tooltip = "$STR_KPLIB_DIALOG_ADMIN_EXP_TT";
-            onButtonClick = "[] call KPLIB_fnc_adm_exportSave";
+            onButtonClick = "[] call KPLIB_fnc_admin_exportSave";
         };
 
         class KPLIB_ImportButton: KPGUI_PRE_DialogButtonC {
@@ -39,7 +39,7 @@ class KPLIB_admin {
             text = "$STR_KPLIB_DIALOG_ADMIN_IMP";
             y = safeZoneY + safeZoneH * (KP_Y_VAL_C + KP_HEIGTH_TITLE + KP_HEIGTH_BUTTON + 2 * KP_SPACING_Y);
             tooltip = "$STR_KPLIB_DIALOG_ADMIN_IMP_TT";
-            onButtonClick = "[] call KPLIB_fnc_adm_importSave";
+            onButtonClick = "[] call KPLIB_fnc_admin_importSave";
         };
 
         class KPLIB_DeleteButton: KPGUI_PRE_DialogButtonC {
@@ -47,14 +47,14 @@ class KPLIB_admin {
             text = "$STR_KPLIB_DIALOG_ADMIN_DEL";
             y = safeZoneY + safeZoneH * (KP_Y_VAL_C + KP_HEIGTH_TITLE + 2 * KP_HEIGTH_BUTTON + 3 * KP_SPACING_Y);
             tooltip = "$STR_KPLIB_DIALOG_ADMIN_DEL_TT";
-            onButtonClick = "[] call KPLIB_fnc_adm_deleteExport";
+            onButtonClick = "[] call KPLIB_fnc_admin_deleteExport";
         };
 
         class KPLIB_WipeButton: KPGUI_PRE_DialogButtonC {
             text = "$STR_KPLIB_DIALOG_ADMIN_WIPE";
             y = safeZoneY + safeZoneH * (KP_Y_VAL_C + KP_HEIGTH_TITLE + 3 * KP_HEIGTH_BUTTON + 4 * KP_SPACING_Y);
             tooltip = "$STR_KPLIB_DIALOG_ADMIN_WIPE_TT";
-            onButtonClick = "[] call KPLIB_fnc_adm_wipe";
+            onButtonClick = "[] call KPLIB_fnc_admin_wipe";
         };
 
         class KPLIB_DialogCross: KPGUI_PRE_DialogCrossC {};
